adds a raster/condition/db testcase to reduce repeated logic in tests
